Some more changes

+++ src/com/thomas/skyrim/tanning/algorithm/CoveredTriangle.java
@@ -5,26 +5,22 @@ import com.thomas.skyrim.tanning.mesh.data.Triangle;
 import com.thomas.skyrim.tanning.mesh.data.Tuple;
 
 import java.awt.*;
-import java.util.List;
 
 
 /**
  *
  */
 public class CoveredTriangle {
-    //TODO -make this remember it's neighbours
-    //TODO -make this paint itself with a gradient. Starting from it's covered value in the center
-    //TODO  to the mean covered value at the edges with the neighbours.
     private final Triangle triangle;
-    private final float covered;
+    private final double covered;
 
-    public CoveredTriangle(Triangle triangle, float covered) {
+    public CoveredTriangle(Triangle triangle, double covered) {
         this.triangle = triangle;
         this.covered = covered;
     }
 
     public void write(Graphics2D graphics, int size) {
-        graphics.setColor(new Color(covered, covered, covered));
+        graphics.setColor(new Color((float) covered, (float) covered, (float) covered));
         Polygon polygon = new Polygon();
         for (Node node : triangle.getNodes()) {
             Tuple uv = node.getUv();
@@ -37,7 +33,7 @@ public class CoveredTriangle {
         return triangle;
     }
 
-    public float getCovered() {
+    public double getCovered() {
         return covered;
     }
